requirements_kimba.txt
PyQt5
pyautogui
pygetwindow
psutil
openai
requests
tqdm
gradio
huggingface_hub
pydantic
uvicorn
fastapi
typing-extensions
markdown-it-py
jinja2
anyio
numpy
colorama
beautifulsoup4
gitpython
lxml
Pillow
pygame
pyttsx3
speechrecognition
tk
torch
transformers
watchdog
# Optional empfehlenswert
python-dotenv
setuptools
wheel
playsound==1.2.2

pyaudio
nest_asyncio
textstat
langdetect
pyyaml
autopip
fuzzywuzzy	
nltk
aiofiles
httpx
orjson
rich
python-multipart


